+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [UNRELEASED] neptune 1.1.0
 
 ### Features
+- Added ability to provide repository path with `GitRef` to `init_run` ([#1292](https://github.com/neptune-ai/neptune-client/pull/1292))
 - `Run`, `Model`, `ModelVersion` and `Project` could be created with constructor in addition to `init_*` functions ([#1246](https://github.com/neptune-ai/neptune-client/pull/1246))
 
 ## neptune 1.0.2